Rebuild querybuilder for indexes and foreign keys
Replace EscapeTableName with a method which uses a RDBMS specific
identifier escape character. Escape character is modifiable which is
useful when the RDBMS is in a non-default ISO compliance mode.

All identifiers are escaped in the QueryBuilders
Add culture invariance to QueryBuilders

Note: SQLite doesn't allow a composite primary key when using auto
      increment, doing so will result in a syntax error.

Signed-off-by: Beau Hastings <beausy@gmail.com>