Removed unimportant messages from OnSecondUpdate()
Removed check ignores failed and player disabled for
not being logged in whilst SSC is enabled console messages, as
they are incessant in larger servers.

The messages are not informative as they are not reasonably
actionable by the console, and there is no point in spamming
server operators about such issues.