Kick players for username with invalid characters can now be toggled
Also added a little not to TShock.cs about win 10 and potenial need for more RAM as win 10 eats a lot (2.6/4GB when doing nothing!). So on windows 10 you need about 4GB of RAM - and that is the minimum for windows my testing is not very in depth so I would look into it more.