@Patrikkk Packet 55 checkpoint
As of this commit, the 1.4.0.2 prep work has been completed. This means
that GetDataHandlers & Bouncer have been updated to support changes to
the protocol made up to packet 55. No exploits have been checked for in
TileSend according to him, but this means that subsequent work can start
on packet 56 and work forward.