Bouncer OnTileEdit - Fix valid replacment action getting caught.
In a previous commit I've added the checks for TileReplace where I believed it was necessary.  By placing a check for the tile replace action, in another check this action got falsly catched and ignored by the player.
This commit fixes that.
I've also moved the block that checks for max valid edit data closer to the start of the method, where we have a check for edit data not being smaller than 0.