Support silent teleport commands
This commit changes the behavior of .tphere and .tp. Specifically, when
teleporting other players, those players will not be told if they've
been teleported silently. The previous behavior of this code path/set of
code paths was that a silent teleport would inform the player that they
were teleported, but it wouldn't tell them who initiated the teleport.

Because the intent of the silent command system is ambiguous, and
because teleporting will always be obvious, I've decided to make the
decision to remove the message/notice entirely. It's possible that
people are using this as an "API" of sorts by teleporting people with
commands in response to events, and thus, the silence makes more sense
in that context. In the context where the person initiating the teleport
doesn't want to be known, though, this can be slightly less clear. I'm
not fully certain if this is the right move, but I figure this is the
best way to update these commands to be silent without adding some new
permission called "Permissions.tpninja" or something.