Add NuSpec file
*** Experimental ***

As of this commit this NuSpec file is hardcoded to use a sha on gen-dev.

To build/publish:

1. Build TSAPI and TShock locally.
2. `mkdir -p ./package/`
3. `cp TShockAPI/bin/Debug/TShockAPI.dll ./package/lib/net451`
4. `cp TShockAPI/bin/Debug/TShockAPI.XML ./package/lib/net451`
5. `cp TShockAPI/bin/Debug/TShockAPI.pdb ./package/lib/net451`
6. `mono ~/bin/nuget.exe pack TShockAPI/tshock.nuspec -BasePath package`

Now you have a package. To publish, log into GitHub:

    dotnet nuget add source --username urname --password \
    a_github_PAT_with_package_publish_scope \
    --store-password-in-clear-text --name github \
    "https://nuget.pkg.github.com/Pryaxis/index.json"

Next, push:

    dotnet push yourpackage.nupkg --api-key the_PAT_from_before \
    --source "github"

If you didn't get a BadRequest error it means that it's been pushed and
is going to slowly propagate out.

Weird quirks here: This is apparently pointed to a github sha but
definitely still versioned like a normal nuget package. So, no idea how
this will work in practice.

Also, GitHub documentation is terrible about how this works. No idea if
authentication is required to install packages or not. No idea if you
can add this stuff to VS directly or not. No real idea how any of this
works. Also, yes, these instructions are partially dotnet and partially
nuget, and heavily based off of @DeathCradle's workflow for OTAPI.