Add Obsidian vault for docs rewrite
Because Readme.io has been kinda scummy in terms of "giving us the
features that we want" and because most wikis don't work well, I'm
proposing a hybrid approach: an Obsidian vault (which is plain markdown)
and having the contents published to Obsidian publish. The end result
will be nice a nice and easily navigable website that can be used to
reference TShock things without needing a separate interface to edit the
docs for.

Plus we can integrate with CI, etc.