Fix firework command failing without color
After updates to the firework command, a guard was lost, allowing the
switch statement that selected the color to execute on a parameter that
was empty. A guard was added to prevent falling into the switch
statement without enough arguments to match a color.

The default type of firework was manually set to a red firework.

Fixes #2507