Swap 0 with PlayerAnnounceResult.None
In 44034c76496fab809d1c5aaccf586b95a087456d I added a magic 0 instead of
a call to an enum because I was looking at an outdated diff that didn't
have `None` as an option. Turns out this was added a day later in a
commit I didn't see originally. Changed to `None` to make code better
documentation wise.