More correctly implement GPLv3 compliance
The previous startup disclaimer probably didn't do a good-enough job of
explaining license terms to people. This updates it emphasizing that
there is no warranty with the software, and that the actual license is
GPL v3 or later, not GPL v3. Further, since we don't ship the license
with the software (yet), this gives people a link to the license. The
license itself says we should provide a link if we don't, so yeah.

Inspired by the fact that we had another user pop into discord asking
why all of their stuff was deleted.

I tried to keep the startup disclaimer roughly-the-same-size as the
original disclaimer (no more than 3 lines, some-ish length), because
this is so annoying but I feel like we actually need to try to be
compliant here.

I'm 60% sure we should implement a command that explains this but I
really don't want to do that.