Fix #2674 - Change the use of current directory to match the server assembly location
This will allow dependencies to be resolved beside the assembly as intended when startup scripts are used outside the working directory.
I am not sure why or how many people would do this, as TSAPI wont load plugins (existing issue), but this is a step in the direction of opening up to allow considerations for that.