Allow the Axe of Regrowth and the Rubblemaker to pass Bouncer checks
The Axe of Regrowth will place a `Saplings` tile object where a tree
used to be, whilst it's `createItem` does not match that of a sapling.
Added an exception to allow this interaction to succeed.

The Rubblemaker is allowed to place echo piles, which will not match
the `createTile` or `placeStyle` of the created piles (except in one
very specific case, I suppose). Added an exception to allow this
interaction to succeed.

The check for the `style` of the placement to match that of
`SelectedItem` was moved to be later on, after checking the tile type,
and only if the two exceptions above weren't triggered. It will also now
sync the tiles upon failure.