Only restrict CC'd players during login whilst SSC is enabled
The player being CC'd does not cause any issues for us if SSC isn't
enabled, so we can relax it's restriction.

This also allows the `RequireLogin` config option to function usefully
again, as it would disable the player, causing them to be CC'd until
they logged in, creating a vicious cycle.