Update TSPlayer.cs
Added the ability to select the cause of death in KillPlayer and DamagePlayer.

Added an xml to send a message in SetPvP

Since GodMode can now be set in Journey mode, we can remove the unnecessary field and add a property whereby we can change the value of GodMode for the player.

Added a property to get the RemoteClient of the player.

Added the ability to get a player's position through a property.

Added an option to update section visibility.